refactor: extract common version helpers
Move shared version parsing, version-distance, and nix package-name
normalization helpers out of common.utils into focused common
modules. Keep compatibility re-exports in common.utils while
switching the remaining Repology, nixupdate, and vulnxscan
consumers over to direct imports.

Add focused tests for the extracted helpers and their
compatibility re-exports so the next common.utils cleanup can
remove more legacy imports with lower risk.
