+++ src/common/utils.py
@@ -10,15 +10,16 @@ import pathlib
 import re
 import subprocess
 
-import packaging.version
-import pandas as pd
+import pandas as _pandas
 
 from common import df as _df
 from common import errors as _errors
 from common import http as _http
 from common import log as _log
 from common import nix_utils as _nix_utils
+from common import package_names as _package_names
 from common import proc as _proc
+from common import versioning as _versioning
 
 ###############################################################################
 
@@ -45,9 +46,14 @@ df_to_csv_file = _df.df_to_csv_file
 exec_cmd = _proc.exec_cmd
 nix_cmd = _proc.nix_cmd
 which = _proc.which
+pd = _pandas
 get_nix_store_dir = _nix_utils.get_nix_store_dir
 normalize_nix_store_path = _nix_utils.normalize_nix_store_path
 parse_nix_derivation_show = _nix_utils.parse_nix_derivation_show
+number_distance = _versioning.number_distance
+version_distance = _versioning.version_distance
+parse_version = _versioning.parse_version
+nix_to_repology_pkg_name = _package_names.nix_to_repology_pkg_name
 
 
 def exit_unless_command_exists(name):
@@ -138,114 +144,6 @@ def get_py_pkg_version(package="sbomnix"):
     except importlib.metadata.PackageNotFoundError:
         versionstr = "0.0.0"
     return versionstr
-
-
-def number_distance(n1, n2):
-    """
-    Return float value between [0.0,1.0] indicating the distance
-    between two non-negative numbers.
-    Returns 1.0 if the two numbers are equal.
-    Returns 0.0 if either argument is not a non-negative number.
-    """
-    if (
-        not isinstance(n1, (float, int))
-        or not isinstance(n2, (float, int))
-        or n1 < 0
-        or n2 < 0
-    ):
-        return 0.0
-    min_n = min(n1, n2)
-    max_n = max(n1, n2)
-    if max_n == 0:
-        return 1.0
-    if min_n == 0:
-        min_n += 1
-        max_n += 1
-    return min_n / max_n
-
-
-def version_distance(v1, v2):
-    """
-    Return float value between [0.0,1.0] indicating the closeness
-    of the given two version number strings.
-    """
-    v1 = str(v1)
-    v2 = str(v2)
-    v1_clean = re.sub(r"[^0-9.]+", "", v1)
-    v2_clean = re.sub(r"[^0-9.]+", "", v2)
-    re_vsplit = re.compile(r".*?(?P<ver_beg>[0-9][0-9]*)(?P<ver_end>.*)$")
-    match = re.match(re_vsplit, v1_clean)
-    if not match:
-        LOG.debug("Unexpected v1 version '%s'", v1)
-        return 0.0
-    v1_major = match.group("ver_beg")
-    v1_minor = match.group("ver_end").replace(".", "")
-    v1_float = float(v1_major + "." + v1_minor)
-    match = re.match(re_vsplit, v2_clean)
-    if not match:
-        LOG.debug("Unexpected v2 version '%s'", v2)
-        return 0.0
-    v2_major = match.group("ver_beg")
-    v2_minor = match.group("ver_end").replace(".", "")
-    v2_float = float(v2_major + "." + v2_minor)
-    return number_distance(v1_float, v2_float)
-
-
-def parse_version(ver_str):
-    """
-    Return comparable version object from the given version string.
-    Returns None if the version string can not be converted to version object.
-    """
-    ver_str = str(ver_str)
-    if not ver_str:
-        return None
-    re_ver = re.compile(r".*?(?P<ver_beg>[0-9][0-9.]*)(?P<ver_end>.*)$")
-    match = re_ver.match(ver_str)
-    if not match:
-        LOG.debug("Unable to parse version '%s'", ver_str)
-        return None
-    ver_beg = match.group("ver_beg").rstrip(".")
-    ver_end = match.group("ver_end")
-    ver_end = re.sub(r"[^0-9.]+", "", ver_end).lstrip(".")
-    if ver_end:
-        ver_end = f"+{ver_end}"
-    else:
-        ver_end = ""
-    ver_end = ver_end.rstrip(".")
-    ver = f"{ver_beg}{ver_end}"
-    ver = re.sub(r"\.+", ".", ver)
-    LOG.log(LOG_SPAM, "%s --> %s", ver_str, ver)
-    if not ver:
-        LOG.debug("Invalid version '%s'", ver_str)
-        return None
-    return packaging.version.parse(ver)
-
-
-def nix_to_repology_pkg_name(nix_pkg_name):
-    """Convert nix package name to repology package name"""
-    if not nix_pkg_name or pd.isnull(nix_pkg_name):
-        return nix_pkg_name
-    # Convert nix_pkg_name so it matches repology package name
-    nix_pkg_name = nix_pkg_name.lower()
-    re_nix_to_repo = re.compile(
-        r"^(?:"
-        r"(python)|(perl)|(emacs)|(vim)plugin|(ocaml)|"
-        r"(gnome)-shell-extension|(lisp)|(ruby)|(lua)|"
-        r"(php)[0-9]*Packages|(go)|(coq)|(rust)"
-        r")"
-        r"[0-9.]*-(.+)"
-    )
-    match = re.match(re_nix_to_repo, nix_pkg_name)
-    if match:
-        # Filter out all non-matched groups
-        matches = list(filter(None, match.groups()))
-        assert len(matches) == 2, f"Unexpected package name '{nix_pkg_name}'"
-        nix_pkg_name = f"{matches[0]}:{matches[1]}"
-    if nix_pkg_name == "python3":
-        nix_pkg_name = "python"
-    if nix_pkg_name == "libtiff":
-        nix_pkg_name = "tiff"
-    return nix_pkg_name
 
 
 def check_positive(val):
